requirements.txt
fastapi==0.135.2
langchain_community==0.4.1
langchain_core==1.2.21
langchain_google_genai==4.2.1
langchain_pinecone==0.2.13
langchain_text_splitters==1.1.1
protobuf==7.34.1
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.32.5
streamlit==1.55.0
